+++ sdk/src/main/java/com/mapbox/maps/renderer/MapboxRenderer.kt
@@ -44,6 +44,12 @@ internal abstract class MapboxRenderer : MapClient {
     }
   }
 
+  internal var snapshotLegacyModeEnabled = false
+    @MainThread
+    get
+    @MainThread
+    set
+
   @UiThread
   fun onDestroy() {
     widgetRenderer.cleanUpAllWidgets()
@@ -144,13 +150,14 @@ internal abstract class MapboxRenderer : MapClient {
     }
     val lock = ReentrantLock()
     val waitCondition = lock.newCondition()
+    val legacyMode = snapshotLegacyModeEnabled
     lock.withLock {
       var snapshot: Bitmap? = null
       renderThread.queueRenderEvent(
         RenderEvent(
           runnable = {
             lock.withLock {
-              snapshot = performSnapshot()
+              snapshot = performSnapshot(legacyMode)
               waitCondition.signal()
             }
           },
@@ -169,9 +176,10 @@ internal abstract class MapboxRenderer : MapClient {
       logE(TAG, "Could not take map snapshot because map is not ready yet.")
       listener.onSnapshotReady(null)
     }
+    val legacyMode = snapshotLegacyModeEnabled
     renderThread.queueRenderEvent(
       RenderEvent(
-        runnable = { listener.onSnapshotReady(performSnapshot()) },
+        runnable = { listener.onSnapshotReady(performSnapshot(legacyMode)) },
         needRender = true,
         eventType = EventType.DEFAULT
       )
@@ -194,15 +202,15 @@ internal abstract class MapboxRenderer : MapClient {
   }
 
   @WorkerThread
-  private fun performSnapshot(): Bitmap? {
+  private fun performSnapshot(legacyMode: Boolean): Bitmap? {
     if (width == 0 && height == 0) {
       logE(TAG, "Could not take map snapshot because map is not ready yet.")
       return null
     }
 
-    if (pixelReader == null || pixelReader?.width != width || pixelReader?.height != height) {
+    if (pixelReader == null || pixelReader?.width != width || pixelReader?.height != height || pixelReader?.legacyMode != legacyMode) {
       pixelReader?.release()
-      pixelReader = PixelReader(width, height)
+      pixelReader = PixelReader(width, height, legacyMode)
     }
     val pixelReader = pixelReader!!
 
@@ -231,15 +239,15 @@ internal abstract class MapboxRenderer : MapClient {
       }
     } catch (e: Throwable) {
       logW(TAG, "Exception ${e.localizedMessage} happened when reading pixels")
-      if (pixelReader.supportsPbo) {
+      if (!pixelReader.legacyMode) {
         logW(TAG, "Re-creating PixelReader with no PBO support and making snapshot again")
         pixelReader.release()
         this.pixelReader = PixelReader(
           width = pixelReader.width,
           height = pixelReader.height,
-          supportsPbo = false
+          legacyMode = true
         )
-        return performSnapshot()
+        return performSnapshot(legacyMode = true)
       }
     }
     return null